add 64bit compat back to it

+++ game/client/c_baseanimating.cpp
@@ -961,7 +961,7 @@ void C_BaseAnimating::LockStudioHdr()
 	
 	if ( pNewWrapper->GetVirtualModel() )
 	{
-		MDLHandle_t hVirtualModel = (MDLHandle_t)(int)(pStudioHdr->VirtualModel())&0xffff;
+		MDLHandle_t hVirtualModel = (MDLHandle_t)(intp)(pStudioHdr->VirtualModel())&0xffff;
 		mdlcache->LockStudioHdr( hVirtualModel );
 	}
 
@@ -993,7 +993,7 @@ void C_BaseAnimating::UnlockStudioHdr()
 			// Immediate-mode rendering, can unlock immediately
 			if ( pStudioHdr->GetVirtualModel() )
 			{
-				MDLHandle_t hVirtualModel = (MDLHandle_t)(int)(pStudioHdr->VirtualModel())&0xffff;
+				MDLHandle_t hVirtualModel = (MDLHandle_t)(intp)(pStudioHdr->VirtualModel())&0xffff;
 				mdlcache->UnlockStudioHdr( hVirtualModel );
 			}
 			mdlcache->UnlockStudioHdr( m_hStudioHdr );
